+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.8 ] - 2024/03/15
+
+  * Added service `playlist_items_add` to add one or more items to a user's playlist.  Items are added in the order they are listed in the `uris` argument.
+
 ###### [ 1.0.7 ] - 2024/03/07
 
   * Updated service `service_spotify_player_media_play_context` to pause the Spotify Connect device before switching play context, and resuming after.